flink-project/requirements.txt
alpaca==1.0.0
alpaca-py==0.34.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
six==1.17.0
sseclient-py==1.8.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==14.1
